wrap bitset util within FixedBitSet to get test
